LayoutTests/fast/dom/shadow/has-content-elements.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="resources/polyfill.js"></script>
<script>
if (!window.internals)
    debug("windows.internals not found!");
else if (!window.internals.hasContentElement)
    debug("windows.internals.hasContentElement not found!");
</script>
</head>
<body>

<p>Test for Bug 100921: <a href="https://bugs.webkit.org/show_bug.cgi?id=100921">https://bugs.webkit.org/show_bug.cgi?id=100921</a> - [Shadow] ShadowRoot should be able to know the existence of &lt;content&gt;</p>

<div id="host"></div>
<pre id="console"></pre>

<script>
var host = document.getElementById("host");
var shadowRoot = host.webkitCreateShadowRoot();

shouldBeFalse("internals.hasContentElement(shadowRoot)");

var div = document.createElement("div");
div.appendChild(document.createElement("content"));
shadowRoot.appendChild(div);
shouldBeTrue("internals.hasContentElement(shadowRoot)");

shadowRoot.removeChild(div);
shouldBeFalse("internals.hasContentElement(shadowRoot)");

shadowRoot.innerHTML = "<div>No Insertion Points</div>";
shouldBeFalse("internals.hasContentElement(shadowRoot)");

shadowRoot.innerHTML = "<content></content>";
shouldBeTrue("internals.hasContentElement(shadowRoot)");

shadowRoot.innerHTML = "<shadow></shadow>";
shouldBeFalse("internals.hasContentElement(shadowRoot)");

shadowRoot.innerHTML = "<div><div><div><content></content></div></div></div>";
shouldBeTrue("internals.hasContentElement(shadowRoot)");

// Case for nested shadow subtree
shadowRoot.innerHTML = '<div>in shadow root</div>';
var innerDiv1 = shadowRoot.querySelector('div');
var innerShadowRoot1 = innerDiv1.webkitCreateShadowRoot();
innerShadowRoot1.innerHTML = "<div><content></content></div>";
var innerDiv2 = innerShadowRoot1.querySelector('div');
var innerShadowRoot2 = innerDiv2.webkitCreateShadowRoot();
innerShadowRoot2.innerHTML = "<div><content title='debugstop'></content></div>";

shouldBeFalse("internals.hasContentElement(shadowRoot)");
shouldBeTrue("internals.hasContentElement(innerShadowRoot1)");
shouldBeTrue("internals.hasContentElement(innerShadowRoot2)");
shadowRoot.innerHTML = "<div></div>";
shouldBeFalse("internals.hasContentElement(shadowRoot)");

// Case for multiple shadow subtrees
shadowRoot.innerHTML = "<div><div><div><content></content></div></div></div>";
var youngerShadowRoot = host.webkitCreateShadowRoot();
// Modifying older shadow root or younger shadow root doesn't affect the other.
shouldBeTrue("internals.hasContentElement(shadowRoot)");
shouldBeFalse("internals.hasContentElement(youngerShadowRoot)");

shadowRoot.innerHTML = "<shadow></shadow>";
shouldBeFalse("internals.hasContentElement(shadowRoot)");
shouldBeFalse("internals.hasContentElement(youngerShadowRoot)");

youngerShadowRoot.innerHTML = "<content></content>";
shouldBeFalse("internals.hasContentElement(shadowRoot)");
shouldBeTrue("internals.hasContentElement(youngerShadowRoot)");
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
